Calculate and display balance, transactions, username
A function was created to calculate "movements" a.k.a transactions for a total balance using a reduce method. It then was displayed in the UI. Transactions can also be seen as withdrawals and deposits with a transaction number.  Simply created with a template literal and inserting adjacent HTML. A function was created to create usernames for the accounts using a forEach method to lowerCase all letters, split each name into arrays, and map and return the beginning of each array to finally .join the arrays back together to create the username.