Display UI, clear Input fields, display movements, display balance, display summary features added
A function was created to listen to an event of click on the login button that would then call a function to find the user account if it exists. If it does, then display all UI messages. Else, an error would be thrown. Rates were changed to match each account uniquely.